Refactor datafetch (fix statistics polling for p > 1)
